requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.4.3
antlr4-python3-runtime==4.9.3
asttokens
async-timeout==4.0.3
attrs==23.2.0
backcall 
beartype==0.17.2
beautifulsoup4==4.12.3
cachetools==5.3.1
causal-conv1d==1.3.0.post1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
comm 
contourpy==1.1.0
coolgpus==0.23
cycler==0.11.0
debugpy 
decorator 
dominate==2.9.1
easydict==1.11
edit-distance==1.0.6
einops==0.7.0
ema-pytorch==0.4.3
entrypoints 
executing 
fastdtw==0.3.4
filelock==3.12.2
fonttools==4.42.0
frozenlist==1.4.1
fsspec==2023.6.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
huggingface-hub==0.23.4
idna==3.4
iisignature 
imagecorruptions==1.1.2
imageio==2.34.0
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel 
ipython 
jedi 
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jupyter-client 
jupyter_core 
kiwisolver==1.4.4
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
lightning-utilities==0.11.2
llvmlite==0.40.1
lmdb==1.4.1
mamba-ssm==2.1.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline 
multidict==6.0.5
natsort==8.4.0
nest_asyncio 
networkx==3.1
ninja==1.11.1.1
numba==0.57.1
numpy==1.24.4
nvidia-ml-py==12.535.77
nvitop==1.2.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.1
pandas==2.0.3
parso 
pexpect 
pickleshare 
Pillow==10.0.0
platformdirs
pretty-errors==1.2.25
prompt-toolkit
protobuf==5.26.0
psutil
ptflops==0.7
ptyprocess
pure-eval
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.0.0
pytorch-metric-learning==2.3.0
pytorch-wavelets==1.3.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.3
six
soupsieve==2.6
stack-data
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.3.2
tokenizers==0.19.1
torchmetrics==1.4.0
torchnet==0.0.4
torchstat==0.0.7
tornado
tqdm==4.65.0
traitlets
transformers==4.41.2
tslearn==0.6.1
typing_extensions
tzdata==2023.3
urllib3==2.0.4
visdom==0.2.4
wcwidth 
websocket-client==1.7.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.16.2
